refactor: move routing file to container folder
